address feedback and upgrade rc-slider to get types
